requirements.txt
discord.py
dotenv
os